Move logic to child
